changed to 16 bit

+++ firmware/src/cardlist.h
@@ -13,14 +13,14 @@ class listElement {
 		bool compare_uid(byte* uid, uint8_t size);
 		byte* get_uuid();
 		uint8_t get_uuidLength();
-		uint8_t get_mode();
-		uint8_t get_folder();
+		uint16_t get_mode();
+		uint16_t get_folder();
 		uint16_t get_track();
-		uint8_t* get_mode_p();
-		uint8_t* get_folder_p();
+		uint16_t* get_mode_p();
+		uint16_t* get_folder_p();
 		uint16_t* get_track_p();
-		void set_mode(uint8_t mode);
-		void set_folder(uint8_t folder);
+		void set_mode(uint16_t mode);
+		void set_folder(uint16_t folder);
 		void set_track(uint16_t track);
 		void set_uuid(byte* uid, uint8_t size);
 		void set_next(listElement* element);
@@ -28,9 +28,9 @@ class listElement {
 		byte		m_uidByte[10];
 		uint8_t	m_uidByteLength;
 		listElement* m_next;
-		uint8_t m_folder;
+		uint16_t m_folder;
 		uint16_t m_track;
-		uint8_t m_mode;
+		uint16_t m_mode;
 };
